Design changes. Added scoring. Some linking between pages added. Changed level definition. Fixed DB.
Changed the appearance of automatons. Added a title bar and links and icons for back (to level list)
and tutorial (to be written). Implemented scoring on successful game and storing the score into the
database. Fixed a broken FOREIGN KEY reference (again). Level definition now contains par scores,
which determine under which numbers for number of transitions, steps and movements player must be
to score. Scoring for number of existing transitions has not yet been implemented.